fix:reset all fragment

+++ app/src/main/java/com/shiv/inventorytracker/ui/DeleteAllFragment.java
@@ -1,5 +1,7 @@
 package com.shiv.inventorytracker.ui;
 
+import android.app.AlertDialog;
+import android.content.DialogInterface;
 import android.os.Bundle;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -25,8 +27,16 @@ public class DeleteAllFragment extends Fragment {
     Button deleteAllBtn = root.findViewById(R.id.deleteAllBtn);
 
     deleteAllBtn.setOnClickListener(v -> {
-      int rows = databaseHelper.deleteAll();
+      new AlertDialog.Builder(root.getContext())
+              .setTitle("Delete")
+              .setMessage("All the entries will be deleted")
+              .setPositiveButton("Yes", (dialog, which) -> {
+                int rows = databaseHelper.deleteAll();
+                Toasty.success(root.getContext(), rows + " entries deleted", Toasty.LENGTH_SHORT, true).show();
+              }).setNegativeButton("No", null)
+              .show();
     });
+
     return root;
   }
 }
